Refactor code writing
