+++ app/common/helpers/template.ts
@@ -1,6 +1,6 @@
 import { compressSync, decompressSync, strFromU8, strToU8 } from 'fflate';
 import { EditorAttributes } from '../constants';
-import { TemplateNode } from '../models';
+import { TemplateNode } from '../models/elements/templateNode';
 
 export function getTemplateNodeFromElement(element: Element): TemplateNode | undefined {
     const dataOnlookId = element.getAttribute(EditorAttributes.DATA_ONLOOK_ID);
