+++ app/common/models/elements/templateNode.ts
@@ -1,0 +1,17 @@
+export interface TemplateTagPosition {
+    line: number;
+    column: number;
+}
+
+export interface TemplateTag {
+    start: TemplateTagPosition;
+    end: TemplateTagPosition;
+}
+
+export interface TemplateNode {
+    path: string;
+    startTag: TemplateTag;
+    endTag: TemplateTag;
+    commit: string;
+    name?: string;
+}
