+++ app/common/models/index.ts
@@ -1,0 +1,14 @@
+import { TemplateNode } from './elements/templateNode';
+
+export interface WriteStyleParam {
+    selector: string;
+    templateNode: TemplateNode;
+    tailwind: string;
+    codeBlock: string;
+}
+
+export interface StyleCodeDiff {
+    original: string;
+    generated: string;
+    param: WriteStyleParam;
+}
