+++ app/common/models/tunnel.ts
@@ -1,0 +1,4 @@
+export interface TunnelResult {
+    url: string;
+    password: string;
+}
