+++ app/electron/main/code/index.ts
@@ -1,80 +1,131 @@
-import generate from '@babel/generator';
-import { parse } from '@babel/parser';
-import traverse from '@babel/traverse';
-import t from '@babel/types';
-import { twMerge } from 'tailwind-merge';
-import { readBlock } from './files';
-import { CodeResult, WriteStyleParams } from '/common/models';
+import { shell } from 'electron';
+import { readFile, writeFile } from './files';
+import { compareTemplateNodes } from '/common/helpers/template';
+import { StyleCodeDiff } from '/common/models';
+import { TemplateNode } from '/common/models/elements/templateNode';
 
-export async function writeStyle(params: WriteStyleParams[]): Promise<CodeResult[]> {
-    const codeResults: CodeResult[] = [];
-    const generateOptions = { retainLines: true, compact: false };
-
-    for (const param of params) {
-        const code = await readBlock(param.templateNode);
-        const ast = parseJsx(code);
-        const original = removeSemiColonIfApplicable(
-            generate(ast, generateOptions, code).code,
-            code,
-        );
-
-        addClassToAst(ast, param.tailwind);
-
-        const generated = removeSemiColonIfApplicable(
-            generate(ast, generateOptions, code).code,
-            code,
-        );
-        codeResults.push({ original, generated, param });
+export async function readTemplateNodeBlocks(templateNodes: TemplateNode[]): Promise<string[]> {
+    const blocks: string[] = [];
+    for (const templateNode of templateNodes) {
+        const block = await readTemplateNodeBlock(templateNode);
+        blocks.push(block);
     }
-    return codeResults;
+    return blocks;
 }
 
-function removeSemiColonIfApplicable(code: string, original: string) {
-    if (!original.endsWith(';') && code.endsWith(';')) {
-        return code.slice(0, -1);
-    }
-    return code;
-}
+export async function readTemplateNodeBlock(templateNode: TemplateNode): Promise<string> {
+    try {
+        const filePath = templateNode.path;
 
-function parseJsx(code: string) {
-    return parse(code, {
-        plugins: ['typescript', 'jsx'],
-    });
-}
+        const startTag = templateNode.startTag;
+        const startRow = startTag.start.line;
+        const startColumn = startTag.start.column;
 
-function addClassToAst(ast: t.File, className: string) {
-    let processed = false;
-    traverse(ast, {
-        JSXOpeningElement(path) {
-            if (processed) {
-                return;
-            }
-            let classNameAttr = null;
-            path.node.attributes.forEach((attribute) => {
-                if (t.isJSXAttribute(attribute) && attribute.name.name === 'className') {
-                    classNameAttr = attribute;
+        const endTag = templateNode.endTag || startTag;
+        const endRow = endTag.end.line;
+        const endColumn = endTag.end.column;
 
-                    if (t.isStringLiteral(attribute.value)) {
-                        attribute.value.value = twMerge(attribute.value.value, className);
-                    }
-                    // Handle className that is an expression (e.g., cn("class1", className))
-                    else if (
-                        t.isJSXExpressionContainer(attribute.value) &&
-                        t.isCallExpression(attribute.value.expression)
-                    ) {
-                        attribute.value.expression.arguments.push(t.stringLiteral(className));
-                    }
+        const fileContent = await readFile(filePath);
+        const lines = fileContent.split('\n');
+
+        const selectedText = lines
+            .slice(startRow - 1, endRow)
+            .map((line: string, index: number, array: string[]) => {
+                if (index === 0 && array.length === 1) {
+                    // Only one line
+                    return line.substring(startColumn - 1, endColumn);
+                } else if (index === 0) {
+                    // First line of multiple
+                    return line.substring(startColumn - 1);
+                } else if (index === array.length - 1) {
+                    // Last line
+                    return line.substring(0, endColumn);
                 }
-            });
+                // Full lines in between
+                return line;
+            })
+            .join('\n');
 
-            if (!classNameAttr) {
-                const newClassNameAttr = t.jsxAttribute(
-                    t.jsxIdentifier('className'),
-                    t.stringLiteral(className),
-                );
-                path.node.attributes.push(newClassNameAttr);
+        return selectedText;
+    } catch (error: any) {
+        console.error('Error reading range from file:', error);
+        throw error;
+    }
+}
+
+export async function writeStyleCodeDiffs(styleCodeDiffs: StyleCodeDiff[]): Promise<boolean> {
+    try {
+        // Write from bottom to prevent line offset
+        const sortedCodeDiffs = styleCodeDiffs
+            .sort((a, b) => compareTemplateNodes(a.param.templateNode, b.param.templateNode))
+            .toReversed();
+        const files = new Map<string, string>();
+
+        for (const result of sortedCodeDiffs) {
+            let fileContent = files.get(result.param.templateNode.path);
+            if (!fileContent) {
+                fileContent = await readFile(result.param.templateNode.path);
             }
-            processed = true;
-        },
-    });
+
+            const updatedFileContent = await getUpdatedFileContent(
+                result.param.templateNode,
+                result.generated,
+                fileContent,
+            );
+            files.set(result.param.templateNode.path, updatedFileContent);
+        }
+
+        for (const [filePath, content] of files) {
+            await writeFile(filePath, content);
+        }
+    } catch (error: any) {
+        console.error('Error writing range to file:', error);
+        return false;
+    }
+    return true;
+}
+
+export async function getUpdatedFileContent(
+    templateNode: TemplateNode,
+    newBlock: string,
+    fileContent: string,
+): Promise<string> {
+    try {
+        const startTag = templateNode.startTag;
+        const startRow = startTag.start.line;
+        const startColumn = startTag.start.column;
+
+        const endTag = templateNode.endTag || startTag;
+        const endRow = endTag.end.line;
+        const endColumn = endTag.end.column;
+
+        const lines = fileContent.split('\n');
+        const before = lines.slice(0, startRow - 1).join('\n');
+        const after = lines.slice(endRow).join('\n');
+
+        const firstLine = lines[startRow - 1].substring(0, startColumn - 1);
+        const lastLine = lines[endRow - 1].substring(endColumn);
+
+        const newFileContent = [before, firstLine + newBlock + lastLine, after].join('\n');
+        return newFileContent;
+    } catch (error: any) {
+        console.error('Error replacing range in file:', error);
+        throw error;
+    }
+}
+
+export function openInVsCode(templateNode: TemplateNode) {
+    const filePath = templateNode.path;
+    const startTag = templateNode.startTag;
+    const endTag = templateNode.endTag || startTag;
+    let command = `vscode://file/${filePath}`;
+
+    if (startTag && endTag) {
+        const startRow = startTag.start.line;
+        const startColumn = startTag.start.column;
+        const endRow = endTag.end.line;
+        const endColumn = endTag.end.column - 1;
+        command += `:${startRow}:${startColumn}:${endRow}:${endColumn}`;
+    }
+    shell.openExternal(command);
 }
