+++ app/package.json
@@ -102,6 +102,7 @@
         "typescript-eslint": "^7.16.0",
         "vite": "^5.1.5",
         "vite-plugin-electron": "^0.28.4",
-        "vite-plugin-electron-renderer": "^0.14.5"
+        "vite-plugin-electron-renderer": "^0.14.5",
+        "vite-plugin-node-polyfills": "^0.22.0"
     }
 }
